# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AI Vetting System API
  version: 1.0.0
  description: API schema for campaign orchestration and vetting workflows.
paths:
...
  /api/campaigns/{id}/candidates/import/:
    post:
      operationId: campaigns_candidates_import_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this vetting campaign.
        required: true
      tags:
      - campaigns
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VettingCampaign'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VettingCampaign'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VettingCampaign'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VettingCampaign'
          description: ''
  /api/campaigns/{id}/dashboard/:
    get:
      operationId: campaigns_dashboard_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this vetting campaign.
        required: true
      tags:
      - campaigns
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VettingCampaign'
          description: ''
  /api/campaigns/{id}/rubrics/versions/:
    post:
      operationId: campaigns_rubrics_versions_create
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this vetting campaign.
        required: true
      tags:
      - campaigns
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VettingCampaign'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VettingCampaign'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VettingCampaign'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VettingCampaign'
          description: ''
  /api/candidates/:
    get:
      operationId: candidates_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - candidates
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCandidateList'
          description: ''
    post:
      operationId: candidates_create
      tags:
      - candidates
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Candidate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Candidate'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
          description: ''
...
  /api/rubrics/criteria/:
    get:
      operationId: rubrics_criteria_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - rubrics
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRubricCriteriaList'
          description: ''
    post:
      operationId: rubrics_criteria_create
      tags:
      - rubrics
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RubricCriteria'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RubricCriteria'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RubricCriteria'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RubricCriteria'
          description: ''
...
components:
  schemas:
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        full_name:
          type: string
          readOnly: true
        email:
          type: string
          format: email
          maxLength: 254
        phone_number:
          type: string
          maxLength: 20
        preferred_channel:
          $ref: '#/components/schemas/CommunicationChannelEnum'
        consent_recording:
          type: boolean
        consent_ai_processing:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - email
      - first_name
      - full_name
      - id
      - last_name
      - updated_at
...
    CommunicationChannelEnum:
      enum:
      - email
      - sms
      type: string
      description: |-
        * `email` - Email
        * `sms` - SMS
...
    CriteriaTypeEnum:
      enum:
      - document
      - consistency
      - interview
      - behavioral
      - technical
      - custom
      type: string
      description: |-
        * `document` - Document Quality
        * `consistency` - Data Consistency
        * `interview` - Interview Performance
        * `behavioral` - Behavioral Assessment
        * `technical` - Technical Competency
        * `custom` - Custom Criterion
...
    PaginatedCandidateList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
...
    PaginatedRubricCriteriaList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/RubricCriteria'
...
    RubricCriteria:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        rubric:
          type: integer
        name:
          type: string
          maxLength: 200
        description:
          type: string
        criteria_type:
          $ref: '#/components/schemas/CriteriaTypeEnum'
        criteria_type_display:
          type: string
          readOnly: true
        scoring_method:
          $ref: '#/components/schemas/ScoringMethodEnum'
        scoring_method_display:
          type: string
          readOnly: true
        weight:
          type: integer
          maximum: 100
          minimum: 0
          description: Weight within parent rubric or category
        minimum_score:
          type: integer
          maximum: 100
          minimum: 0
          nullable: true
          description: Minimum required score for this criterion
        is_mandatory:
          type: boolean
          description: Must pass this criterion to pass overall
        evaluation_guidelines:
          type: string
          description: Instructions for manual evaluation
        display_order:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
      required:
      - criteria_type
      - criteria_type_display
      - description
      - id
      - name
      - rubric
      - scoring_method_display
      - weight
...
    ScoringMethodEnum:
      enum:
      - ai_score
      - manual_rating
      - binary
      - calculated
      type: string
      description: |-
        * `ai_score` - AI-Generated Score
        * `manual_rating` - Manual Rating
        * `binary` - Pass/Fail
        * `calculated` - Calculated from Sub-Criteria
...
    VettingCampaign:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
        status:
          $ref: '#/components/schemas/VettingCampaignStatusEnum'
        starts_at:
          type: string
          format: date-time
          nullable: true
        ends_at:
          type: string
          format: date-time
          nullable: true
        settings_json: {}
        initiated_by:
          type: integer
          readOnly: true
        initiated_by_email:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - initiated_by
      - initiated_by_email
      - name
      - updated_at
    VettingCampaignStatusEnum:
      enum:
      - draft
      - active
      - closed
      - archived
      type: string
      description: |-
        * `draft` - Draft
        * `active` - Active
        * `closed` - Closed
        * `archived` - Archived
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT